Change common filename and cleanup
