--- cellarium/ml/models/common/fclayer.py
@@ -1,247 +1,0 @@
-# Copyright Contributors to the Cellarium project.
-# SPDX-License-Identifier: BSD-3-Clause
-
-import collections
-from collections.abc import Iterable
-
-import torch
-
-
-class LinearInputBias(torch.nn.Linear):
-    """A `torch.nn.Linear` layer where bias is an input to the forward pass.
-
-    Args:
-        in_features: passed to `torch.nn.Linear`
-        out_features: passed to `torch.nn.Linear`
-        bias: passed to `torch.nn.Linear`
-    """
-
-    def __init__(self, in_features: int, out_features: int, bias: bool = True):
-        super().__init__(in_features, out_features, bias=bias)
-
-    def forward(self, x: torch.Tensor, bias: torch.Tensor) -> torch.Tensor:
-        """
-        Computes the forward pass of the layer as
-        out = x @ self.weight.T + self.bias + bias
-        """
-        return super().forward(x) + bias
-
-
-class DressedLayer(torch.nn.Module):
-    """
-    Small block comprising a `~torch.nn.Module` with optional batch/layer normalization 
-    and configurable activation and dropout.
-
-    Similar to
-    torch.nn.Sequential(
-        layer,
-        optional batch normalization,
-        optional layer normalization,
-        optional activation,
-        optional dropout,
-    )
-    but the `layer` can take multiple inputs.
-
-    Note that batch normalization and layer normalization are mutually exclusive options.
-
-    Args:
-        layer: single layer `torch.nn.Module`, such as an instance of `torch.nn.Linear`
-        use_batch_norm: whether to use batch normalization
-        use_layer_norm: whether to use layer normalization
-        activation_fn: the activation function to use
-        dropout_rate: dropout rate, can be zero
-    """
-
-    def __init__(
-        self,
-        layer: torch.nn.Module,
-        use_batch_norm: bool = False,
-        batch_norm_kwargs: dict = {'momentum': 0.01, 'eps': 0.001},
-        use_layer_norm: bool = False,
-        layer_norm_kwargs: dict = {'elementwise_affine': False},
-        activation_fn: torch.nn.Module | None = torch.nn.ReLU,
-        dropout_rate: float = 0,
-    ):
-        assert not (use_batch_norm and use_layer_norm), "Cannot use both batch and layer normalization."
-        super().__init__()
-        out_features = layer.out_features
-        batch_norm = torch.nn.BatchNorm1d(out_features, **batch_norm_kwargs) if use_batch_norm else None
-        layer_norm = torch.nn.LayerNorm(out_features, **layer_norm_kwargs) if use_layer_norm else None
-        activation = activation_fn() if (activation_fn is not None) else None
-        dropout = torch.nn.Dropout(p=dropout_rate) if dropout_rate > 0 else None
-        module_list = [batch_norm, layer_norm, activation, dropout]
-        self.layer = layer
-        self.dressing = torch.nn.Sequential(*[m for m in module_list if m is not None])
-
-    def forward(self, *args, **kwargs) -> torch.Tensor:
-        """
-        Computes the forward pass of the block.
-        """
-        x = self.layer(*args, **kwargs)
-        return self.dressing(x)
-
-
-class FCLayers(torch.nn.Module):
-    """A helper class to build fully-connected layers for a neural network.
-
-    Parameters
-    ----------
-    n_in
-        The dimensionality of the input
-    n_out
-        The dimensionality of the output
-    n_cat_list
-        A list containing, for each category of interest,
-        the number of categories. Each category will be
-        included using a one-hot encoding.
-    n_layers
-        The number of fully-connected hidden layers
-    n_hidden
-        The number of nodes per hidden layer
-    dropout_rate
-        Dropout rate to apply to each of the hidden layers
-    use_batch_norm
-        Whether to have `BatchNorm` layers or not
-    use_layer_norm
-        Whether to have `LayerNorm` layers or not
-    use_activation
-        Whether to have layer activation or not
-    bias
-        Whether to learn bias in linear layers or not
-    inject_covariates
-        Whether to inject covariates in each layer, or just the first (default).
-    activation_fn
-        Which activation function to use
-    """
-
-    def __init__(
-        self,
-        n_in: int,
-        n_out: int,
-        n_cat_list: Iterable[int] = None,
-        n_layers: int = 1,
-        n_hidden: int = 128,
-        dropout_rate: float = 0.1,
-        use_batch_norm: bool = True,
-        use_layer_norm: bool = False,
-        use_activation: bool = True,
-        bias: bool = True,
-        inject_covariates: bool = True,
-        activation_fn: torch.nn.Module = torch.nn.ReLU,
-    ):
-        super().__init__()
-        self.inject_covariates = inject_covariates
-        layers_dim = [n_in] + (n_layers - 1) * [n_hidden] + [n_out]
-
-        if n_cat_list is not None:
-            # n_cat = 1 will be ignored
-            self.n_cat_list = [n_cat if n_cat > 1 else 0 for n_cat in n_cat_list]
-        else:
-            self.n_cat_list = []
-
-        cat_dim = sum(self.n_cat_list)
-        self.fc_layers = torch.nn.Sequential(
-            collections.OrderedDict(
-                [
-                    (
-                        f"Layer {i}",
-                        torch.nn.Sequential(
-                            torch.nn.Linear(
-                                n_in + cat_dim * self.inject_into_layer(i),
-                                n_out,
-                                bias=bias,
-                            ),
-                            # non-default params come from defaults in original Tensorflow implementation
-                            torch.nn.BatchNorm1d(n_out, momentum=0.01, eps=0.001) if use_batch_norm else None,
-                            torch.nn.LayerNorm(n_out, elementwise_affine=False) if use_layer_norm else None,
-                            activation_fn() if use_activation else None,
-                            torch.nn.Dropout(p=dropout_rate) if dropout_rate > 0 else None,
-                        ),
-                    )
-                    for i, (n_in, n_out) in enumerate(zip(layers_dim[:-1], layers_dim[1:]))
-                ]
-            )
-        )
-
-    def inject_into_layer(self, layer_num) -> bool:
-        """Helper to determine if covariates should be injected."""
-        user_cond = layer_num == 0 or (layer_num > 0 and self.inject_covariates)
-        return user_cond
-
-    def set_online_update_hooks(self, hook_first_layer=True):
-        """Set online update hooks."""
-        self.hooks = []
-
-        def _hook_fn_weight(grad):
-            categorical_dims = sum(self.n_cat_list)
-            new_grad = torch.zeros_like(grad)
-            if categorical_dims > 0:
-                new_grad[:, -categorical_dims:] = grad[:, -categorical_dims:]
-            return new_grad
-
-        def _hook_fn_zero_out(grad):
-            return grad * 0
-
-        for i, layers in enumerate(self.fc_layers):
-            for layer in layers:
-                if i == 0 and not hook_first_layer:
-                    continue
-                if isinstance(layer, torch.nn.Linear):
-                    if self.inject_into_layer(i):
-                        w = layer.weight.register_hook(_hook_fn_weight)
-                    else:
-                        w = layer.weight.register_hook(_hook_fn_zero_out)
-                    self.hooks.append(w)
-                    b = layer.bias.register_hook(_hook_fn_zero_out)
-                    self.hooks.append(b)
-
-    def forward(self, x: torch.Tensor, *cat_list: int):
-        """Forward computation on ``x``.
-
-        Parameters
-        ----------
-        x
-            tensor of values with shape ``(n_in,)``
-        cat_list
-            list of category membership(s) for this sample
-
-        Returns
-        -------
-        :class:`torch.Tensor`
-            tensor of shape ``(n_out,)``
-        """
-        one_hot_cat_list = []  # for generality in this list many indices useless.
-
-        if len(self.n_cat_list) > len(cat_list):
-            raise ValueError("nb. categorical args provided doesn't match init. params.")
-        for n_cat, cat in zip(self.n_cat_list, cat_list):
-            if n_cat and cat is None:
-                raise ValueError("cat not provided while n_cat != 0 in init. params.")
-            if n_cat > 1:  # n_cat = 1 will be ignored - no additional information
-                if cat.ndim not in [1, 2]:
-                    raise ValueError("Injected categorical cat must be 1D or 2D tensor.")
-                if cat.ndim == 1:
-                    one_hot_cat = torch.nn.functional.one_hot(cat.long(), n_cat).float()
-                else:
-                    assert cat.size(1) == n_cat, f"Injected categorical cat was 2D {cat.shape} but the second dim was not n_cat {n_cat}."
-                    one_hot_cat = cat  # cat has already been one_hot encoded
-                one_hot_cat_list += [one_hot_cat]
-        for i, layers in enumerate(self.fc_layers):
-            for layer in layers:
-                if layer is not None:
-                    if isinstance(layer, torch.nn.BatchNorm1d):
-                        if x.dim() == 3:
-                            x = torch.cat([(layer(slice_x)).unsqueeze(0) for slice_x in x], dim=0)
-                        else:
-                            x = layer(x)
-                    else:
-                        if isinstance(layer, torch.nn.Linear) and self.inject_into_layer(i):
-                            if x.dim() == 3:
-                                one_hot_cat_list_layer = [
-                                    o.unsqueeze(0).expand((x.size(0), o.size(0), o.size(1))) for o in one_hot_cat_list
-                                ]
-                            else:
-                                one_hot_cat_list_layer = one_hot_cat_list
-                            x = torch.cat((x, *one_hot_cat_list_layer), dim=-1)
-                        x = layer(x)
-        return x
